requirement.txt
asgiref==3.6.0

certifi==2022.12.7

cffi==1.15.1

charset-normalizer==3.1.0

decouple==0.0.7

Django==4.1.7

djangorestframework==3.14.0

djangorestframework-simplejwt==5.2.2

et-xmlfile==1.1.0

idna==3.4

psycopg2==2.9.6

pycparser==2.21

PyJWT==2.8.0

python-decouple==3.8

python-dotenv==1.0.0

pytz==2023.3

requests==2.28.2

six==1.16.0

sqlparse==0.4.3

tzdata==2022.7

urllib3==1.26.15

XlsxWriter==3.1.